Refactor isBasicAuth to requireBasicAuth
As this middleware is effectively a decisionmaker based on application
mode, it makes more sense to group it with the featureToggle middleware.
