+++ app/src/middleware/featureToggle.js
@@ -1,6 +1,32 @@
 const Problem = require('api-problem');
 const config = require('config');
+
 const log = require('../components/log')(module.filename);
+const { AuthMode, AuthType } = require('../components/constants');
+const { getAppAuthMode } = require('../components/utils');
+
+/**
+ * @function requireBasicAuth
+ * Only allows basic authentication requests if application is in the appropriate mode
+ * @param {object} req Express request object
+ * @param {object} res Express response object
+ * @param {function} next The next callback function
+ * @returns {function} Express middleware function
+ */
+const requireBasicAuth = (req, res, next) => {
+  const authMode = getAppAuthMode();
+  const authType = req.currentUser ? req.currentUser.authType : undefined;
+
+  if (authMode === AuthMode.OIDCAUTH) {
+    return new Problem(501, { detail: 'This action is not supported in the current authentication mode' }).send(res);
+  }
+
+  if (authMode === AuthMode.BASICAUTH || authMode === AuthMode.FULLAUTH && authType !== AuthType.BASIC) {
+    return new Problem(403, { detail: 'User lacks permission to complete this action' }).send(res);
+  }
+
+  next();
+};
 
 /**
  * @function requireDb
@@ -29,5 +55,5 @@ const requireDb = (req, res, next) => {
 };
 
 module.exports = {
-  requireDb
+  requireBasicAuth, requireDb
 };
